[HEIDELPAYSUPPORT-127] Allow plugin to be PHP8.0 compatible and fix eslint issues

See merge request unzerdev/shopware5!27

+++ Components/Hydrator/ResourceHydrator/BasketHydrator.php
@@ -80,7 +80,7 @@ class BasketHydrator implements ResourceHydratorInterface
                 $basketItem->setAmountPerUnit(round($amountPerUnit, self::UNZER_DEFAULT_PRECISION));
                 $basketItem->setAmountGross($amountGross);
                 $basketItem->setAmountNet(round(abs($lineItem['amountnetNumeric']), self::UNZER_DEFAULT_PRECISION));
-                $basketItem->setAmountVat(round(abs(str_replace(',', '.', $lineItem['tax'])), self::UNZER_DEFAULT_PRECISION));
+                $basketItem->setAmountVat(round(abs((float) str_replace(',', '.', $lineItem['tax'])), self::UNZER_DEFAULT_PRECISION));
                 $basketItem->setVat((float) $lineItem['tax_rate']);
             }
